requirements/base.txt
# Base apps
psycopg2
Django
requests
lxml
djangorestframework
